+++ src/components/Stepper/stories/StepperClickProp.template.tsx
@@ -22,10 +22,17 @@ export const StepperClickPropTemplate = () => {
   return (
     <>
       <Stepper activeStep={2} data-testid="stepper">
-        {steps.map(({ content, ...step }) => {
+        {steps.map(({ content, key, completed }) => {
           return (
-            // eslint-disable-next-line no-console
-            <Step {...step} onClick={(step: any) => console.log(`step ${step.index}`)} data-testid="step">
+            <Step
+              completed={completed}
+              key={`step-${key}`}
+              onClick={(step: { index: number; active: boolean; completed: boolean; disabled?: boolean }) => {
+                // Обработчик клика по шагу
+                console.log(`step ${step.index}`);
+              }}
+              data-testid="step"
+            >
               <StepContent tooltipProps={{ style: { maxWidth: '300px' } }}>{content}</StepContent>
             </Step>
           );
@@ -33,10 +40,16 @@ export const StepperClickPropTemplate = () => {
       </Stepper>
       <Separator />
       <Stepper activeStep={2} orientation="vertical" style={{ width: '225px' }}>
-        {steps.map(({ content, ...step }) => {
+        {steps.map(({ content, key, completed }) => {
           return (
-            // eslint-disable-next-line no-console
-            <Step {...step} onClick={(step: any) => console.log(`step ${step.index}`)}>
+            <Step
+              completed={completed}
+              key={`step-${key}`}
+              onClick={(step: { index: number; active: boolean; completed: boolean; disabled?: boolean }) => {
+                // Обработчик клика по шагу
+                console.log(`step ${step.index}`);
+              }}
+            >
               <StepContent tooltipProps={{ style: { maxWidth: '300px' } }}>{content}</StepContent>
             </Step>
           );